refactor: weather component now uses the new useWeather composable
