+++ frontend/src/lib/types/index.ts
@@ -11,13 +11,12 @@ export type {
 // Export weather types
 export type {
   WeatherData,
-  WeatherCondition,
-  WeatherMain,
-  WeatherSys,
   WeatherCache,
-  WeatherServiceConfig,
-  WeatherResult,
   WeatherProps,
+  WeatherState,
+  GeoLocation,
+  IPInfo,
+  LocationData,
   LoadingStateType,
   WeatherErrorTypeType
 } from './weather.js';
@@ -25,39 +24,12 @@ export type {
 export {
   LoadingState,
   WeatherErrorType,
-  ERROR_MESSAGES
+  ERROR_MESSAGES,
+  WEATHER_API_BASE,
+  IP_INFO_API,
+  CACHE_DURATION,
+  GEO_OPTIONS
 } from './weather.js';
-
-// Export location types
-export type {
-  GeoLocation,
-  IPInfo,
-  LocationData,
-  LocationServiceConfig,
-  LocationResult,
-  IPLocationResult,
-  GeolocationOptions
-} from './location.js';
-
-export {
-  DEFAULT_GEO_OPTIONS,
-  DEFAULT_LOCATION_CONFIG,
-  IP_INFO_API
-} from './location.js';
-
-// Export cache types
-export type {
-  CacheEntry,
-  CacheConfig,
-  CacheStats
-} from './cache.js';
-
-export {
-  DEFAULT_CACHE_CONFIG,
-  generateLocationCacheKey,
-  generateWeatherCacheKey,
-  generateIPCacheKey
-} from './cache.js';
 
 // Export debug types (re-exported from weatherDebug.ts)
 export type {
